Changed the partition schedule to be wrapped in a class
