+++ src/kernel/core/scheduling/partitionscheduling.cpp
@@ -4,20 +4,13 @@
 #include <circle/time.h>
 #include <core_schedule.hpp>
 #include <defines.hpp>
-#include <generated_arinc_module.hpp>
 #include <generated_partition_schedule.hpp>
 #include <partition.hpp>
 #include <partition_schedule.hpp>
+#include <process_schedule.hpp>
 
-struct RunningPartition {
-    NAME_TYPE partitionName;
-    APEX_INTEGER startTime;
-    APEX_INTEGER endTime;
-    int index;
-    int partitionAmount;
-};
-
-RunningPartition* getNextPartition(RunningPartition* runningPartition, int size)
+RunningPartition* CyclicExecutiveSchedule::getNextPartition(RunningPartition* runningPartition,
+                                                            int size)
 {
     // Get first partitions
     if (runningPartition[0].endTime == 0) {
@@ -90,7 +83,7 @@ RunningPartition* getNextPartition(RunningPartition* runningPartition, int size)
     return runningPartition;
 }
 
-void partitionScheduler()
+void CyclicExecutiveSchedule::partitionScheduler()
 {
 #if KERNEL_PROCESSER(IS_MULTICORE)
     int size = 4;
@@ -106,14 +99,14 @@ void partitionScheduler()
 #if KERNEL_DEBUG()
     CLogger::Get()->Write("FreeTACOS", LogNotice, "Starting partition schedule");
 #endif
-    getNextPartition(runningPartition, size);
+
     while (1) {
         for (int i = 0; i < size; i++) {
-            if (CTimer::Get()->GetClockTicks() >= runningPartition[i].endTime) {
+            if (runningPartition[0].endTime == 0 ||
+                CTimer::Get()->GetClockTicks() >= runningPartition[i].endTime) {
                 getNextPartition(runningPartition, size);
                 break;
             }
         }
-        // TODO: call process scheduler her
     }
 }
