+++ src/kernel/core/tacoskernel.cpp
@@ -17,6 +17,7 @@ CTacosKernel::CTacosKernel()
 
 CStdlibApp::TShutdownMode CTacosKernel::Run(void)
 {
+    CyclicExecutiveSchedule partitionSchedule;
     mLogger.Write("Tester", LogNotice, "Testing ProcessSchedules..");
     mLogger.Write("ProcessSchedule", LogNotice,
                   "Initialising schedules from XML");
@@ -43,6 +44,7 @@ CStdlibApp::TShutdownMode CTacosKernel::Run(void)
     // Running Entry Process
     auto entry = new Entry(&mLogger);
     entry->Run();
+    partitionSchedule.partitionScheduler();
 
     while (1) {
     }
